fix(options): keep lifecycle under hotspot budget
